clear

+++ migrations/2019-02-20-054109_create_login_history_table/up.sql
@@ -2,6 +2,6 @@
 CREATE TABLE login_history
 (
     id SERIAL PRIMARY KEY NOT NULL,
-    user_id INTEGER NOT NULL REFERENCES users(id),
+    user_id BIGINT NOT NULL REFERENCES users(id),
     login_timestamp TIMESTAMP WITH TIME ZONE NOT NULL
 );
